Update / Repurpose All Enigmatic Technologies for v2.0
Researching the enigmatic encoder and decoder (plus the relevant tier 5 techs) grants countries the option to research the dark matter reactor & deflector and enigmatic destabilisation field.

The event chain associated with this mod is now significantly more complex, with a fair amount of user-visible text explaining what's going on.

The enigmatic disruption field is set up to be equivalent to a shield dampener + quantum destabilizer.