Buy Influence: Various bugfixes & tweaks
Converted trust gain to yearly pulse, to prevent the player from continually starting and cancelling campaigns to get unlimited trust. Unfortunately, the yearly pulse isn't working, for some unknown reason...

Structured the yearly pulse to make it easy to add future flavour events while the campaign is running.

Tweaked the scripted_triggers so that it's easier to find out which type of propaganda campaign is currently active.

Added missing localisation for monthly influence and consumer goods.

Added headings to the purchase response, to remind the player what they clicked on. (It's very tempting to look at the mechanical implications of each buff, rather than the flavour text.)

Changed the buff duration to the production value (10 years).

Ensured that the artist_enclave_diplomacy_engaged flag is cleared in every GOODBYE option. (So the artists don't become unresponsive.)