Buy Influence: Final bugfixes & tweaks before workshop upload
Bugfix: Trust with artisan enclave is increased every year that a propaganda campaign is active, as desired.

Renewal screen shows a reminder of which campaign has just expired.

Custom tooltip explaining trust gain is now embedded in the modifier. (Previously, it was only present in the event.) This means you can view the expected trust gain from the empire screen at any time.